Expand `nix-darwin` setup and update README
An explanation of why se set the tmux `default-shell` option follows
below.

Make sure tmux starts non-login `bash` shells. If this is not specified,
tmux will start by default a *login* shell using the value of the
`default-shell` option (see `man tmux`).

This is problematic with our nix setup, since a login shell will source
`/etc/profile` (appears to be vendored by Apple), and this first runs
`eval $(/usr/libexec/path_helper -s)`. This sets `PATH` to a value where
the nix directories are at the end, meaning nix-provided packages will
not be picked up if they are found in one of the other directories (for
example `git` will use `/usr/bin/git` instead of our nix-provided
`git`). The `/etc/profile` file then goes on to source `/etc/bashrc`
(which is generated by `nix-darwin`).

Normally, when `/etc/bashrc` is sourced as part of the very first login
shell we create (e.g. when launching Alacritty), the
`__NIX_DARWIN_SET_ENVIRONMENT_DONE` environment variable is not set, and
we enter a branch that sources a nix `set-environment` script, that
correctly sets our `PATH` for use with nix. This `set-environment`
script then proceeds to set `__NIX_DARWIN_SET_ENVIRONMENT_DONE`, so that
"child shells [need] not needlessly reset any environment that should be
inherited" [0]. This means the next time a login shell is started (as a
child shell of this first original login shell), the `set-environment`
script is not run again, and `PATH` is left "broken" due to the
manipulations by `/usr/libexec/path_helper` that was done in
`/etc/profile`!

The solution is to just set the tmux `default-command` option, which
will use the value of the option when creating a new window. By
specifying it to just `bash`, a non-login shell will be created (as
opposed to `bash --login` which will create a login shell, or leaving it
unset which means tmux will create a login shell for us, for example).
Non-login shells only source `~/.bashrc`, meaning the shell will not go
through `/etc/profile` that breaks `PATH`.

[0] https://github.com/nix-darwin/nix-darwin/commit/676ef103771aa3fc4b150290294b8ad5610d2750